modules/api/locations/requirements.txt
grpcio==1.42.0
grpcio-tools==1.42.0
SQLAlchemy==1.3.19
GeoAlchemy2==0.8.4
marshmallow==3.7.1
marshmallow-sqlalchemy==0.23.1
psycopg2-binary==2.8.5
shapely==1.7.0